ingest-etl/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
google-cloud-pubsub==2.*
